remove bus arg which prevents some configurations from working in Linux guests
before this change, it worked with an HD 6770 in the 3rd host pci slot, but not with an r7 260x in the 1st slot; and now it works with either GPU
also tested it works with bus=pcie.0 or bus=pcie.0,addr=1.0

and some history... the reason I added bus and addr to some of these lines is some problem I had with usb devices in windows guests. Some windows versions seemed to need them, and others didn't. And also there was different behavior in the qemu boot menu selector depending on how it was done. So I really only cared about it for usb pci controllers, not GPUs. And I don't know if they are still needed anywhere.